refactor: change message_id, user_id, guild_id, channel_id types to u64 in messages and pending_flushes models

+++ entity/src/pending_flushes.rs
@@ -5,14 +5,15 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "pending_flushes")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub message_id: i64,
-    pub notification_id: i64,
-    pub channel_id: i64,
-    pub toilet_id: i64,
-    pub author_id: i64,
-    pub flusher_id: i64,
-    pub threshold_count: i64,
+    #[sea_orm(primary_key, unique, auto_increment = false)]
+    pub message_id: u64,
+    #[sea_orm(unique)]
+    pub notification_id: u64,
+    pub channel_id: u64,
+    pub toilet_id: u64,
+    pub author_id: u64,
+    pub flusher_id: u64,
+    pub threshold_count: u64,
     pub created_at: DateTime<Utc>,
 }
 
@@ -23,23 +24,23 @@ impl ActiveModelBehavior for ActiveModel {}
 
 impl Model {
     pub fn toilet(&self) -> serenity::model::id::ChannelId {
-        serenity::model::id::ChannelId::from(self.toilet_id as u64)
+        serenity::model::id::ChannelId::from(self.toilet_id)
     }
 
     pub fn flusher(&self) -> serenity::model::id::UserId {
-        serenity::model::id::UserId::from(self.flusher_id as u64)
+        serenity::model::id::UserId::from(self.flusher_id)
     }
 
     pub fn message_id(&self) -> serenity::model::id::MessageId {
-        serenity::model::id::MessageId::from(self.message_id as u64)
+        serenity::model::id::MessageId::from(self.message_id)
     }
 
     pub fn notification_id(&self) -> serenity::model::id::MessageId {
-        serenity::model::id::MessageId::from(self.notification_id as u64)
+        serenity::model::id::MessageId::from(self.notification_id)
     }
 
     pub fn channel_id(&self) -> serenity::model::id::ChannelId {
-        serenity::model::id::ChannelId::from(self.channel_id as u64)
+        serenity::model::id::ChannelId::from(self.channel_id)
     }
 
     pub fn threshold(&self) -> u64 {
